Remove internal import from pipeline builder
